+++ core/src/com/unciv/logic/city/City.kt
@@ -234,6 +234,8 @@ class City : IsPartOfGameInfoSerialization {
     internal fun getMaxHealth() =
         200 + cityConstructions.getBuiltBuildings().sumOf { it.cityHealth }
 
+    fun getStrength() = cityConstructions.getBuiltBuildings().sumOf { it.cityStrength }.toFloat() 
+
     override fun toString() = name // for debug
 
     fun isHolyCity(): Boolean = religion.religionThisIsTheHolyCityOf != null && !religion.isBlockedHolyCity
@@ -481,7 +483,7 @@ class City : IsPartOfGameInfoSerialization {
                 + religion.getUniques().filter { it.type == uniqueType }
             ).filter {
                 !it.isTimedTriggerable && it.conditionalsApply(stateForConditionals)
-            }
+            }.flatMap { it.getMultiplied(stateForConditionals) }
     }
 
     // Uniques special to this city
@@ -489,6 +491,7 @@ class City : IsPartOfGameInfoSerialization {
         val uniques = cityConstructions.builtBuildingUniqueMap.getUniques(uniqueType).filter { it.isLocalEffect } +
             religion.getUniques().filter { it.type == uniqueType }
         return if (uniques.any()) uniques.filter { !it.isTimedTriggerable && it.conditionalsApply(stateForConditionals) }
+            .flatMap { it.getMultiplied(stateForConditionals) }
         else uniques
     }
 
@@ -497,7 +500,7 @@ class City : IsPartOfGameInfoSerialization {
         val uniques = cityConstructions.builtBuildingUniqueMap.getUniques(uniqueType)
         // Memory performance showed that this function was very memory intensive, thus we only create the filter if needed
         return if (uniques.any()) uniques.filter { !it.isLocalEffect && !it.isTimedTriggerable
-            && it.conditionalsApply(stateForConditionals) }
+            && it.conditionalsApply(stateForConditionals) }.flatMap { it.getMultiplied(stateForConditionals) }
         else uniques
     }