+++ core/src/com/unciv/ui/screens/civilopediascreen/CivilopediaSearchPopup.kt
@@ -80,7 +80,7 @@ class CivilopediaSearchPopup(
                 val regex = Regex(text, RegexOption.IGNORE_CASE)
                 checkLine = { regex.containsMatchIn(it) }
             } catch (ex: Exception) {
-                ToastPopup("Invalid regular expression", pediaScreen, 4000).open(true)
+                ToastPopup("Invalid regular expression", pediaScreen, 4000)
                 searchButton.enable()
                 return
             }